Update to work with latest master refactorings
