+++ src/app/plugins/kubernetes/components/infrastructure/clusters/bareos/BareOsClusterReviewTable.js
